+++ ChangeLog.md
@@ -1,5 +1,16 @@
 # BagOfStuff ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- `Pipe` can now start with zero arguments.
+  ([#5](https://github.com/davep/bagofstuff/pull/5))
+- Added `Pipe.Nullary` as the type to mark a pipe that takes no arguments.
+  ([#5](https://github.com/davep/bagofstuff/pull/5))
+- An empty `Pipe` when called will now raise a `TypeError`.
+  ([#5](https://github.com/davep/bagofstuff/pull/5))
+
 ## v0.0.1
 
 **Released: 2026-01-31**